[STYLE] 구현체 내용 및 패키지 이동 (user -> auth)
