+++ src/main/java/sopt/makers/authentication/application/auth/api/SocialAccountApi.java
@@ -8,4 +8,6 @@ import org.springframework.http.ResponseEntity;
 public interface SocialAccountApi {
   ResponseEntity<BaseResponse<?>> updateSocialAccount(
       SocialAccountRequest.UpdateSocialAccount socialAccountInfo);
+
+  ResponseEntity<BaseResponse<?>> getRegisterSocialAccountPlatform(String name, String phone);
 }
