+++ src/main/java/sopt/makers/authentication/application/auth/api/SocialAccountApiController.java
@@ -1,15 +1,15 @@
 package sopt.makers.authentication.application.auth.api;
 
 import sopt.makers.authentication.application.auth.dto.request.SocialAccountRequest;
+import sopt.makers.authentication.application.auth.dto.response.AuthResponse;
 import sopt.makers.authentication.support.code.domain.success.SocialAccountSuccess;
 import sopt.makers.authentication.support.common.api.BaseResponse;
 import sopt.makers.authentication.support.util.ResponseUtil;
+import sopt.makers.authentication.usecase.auth.port.in.GetSocialAccountUsecase;
 import sopt.makers.authentication.usecase.auth.port.in.UpdateSocialAccountUsecase;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import lombok.RequiredArgsConstructor;
 
@@ -17,7 +17,26 @@ import lombok.RequiredArgsConstructor;
 @RequestMapping("/api/v1/social/accounts")
 @RequiredArgsConstructor
 public class SocialAccountApiController implements SocialAccountApi {
+  private final GetSocialAccountUsecase getSocialAccountUsecase;
   private final UpdateSocialAccountUsecase updateSocialAccountUsecase;
+
+  @Override
+  @GetMapping("/platform")
+  public ResponseEntity<BaseResponse<?>> getRegisterSocialAccountPlatform(
+      @RequestParam(value = "name", required = false) String name,
+      @RequestParam(value = "phone") String phone) {
+    GetSocialAccountUsecase.GetSocialAccountPlatformCommand command =
+        new GetSocialAccountUsecase.GetSocialAccountPlatformCommand(name, phone);
+    GetSocialAccountUsecase.SocialAccountPlatformInfo socialAccountPlatform =
+        getSocialAccountUsecase.getSocialAccountPlatform(command);
+
+    return ResponseEntity.status(
+            SocialAccountSuccess.GET_SOCIAL_ACCOUNT_PLATFORM.getStatus().value())
+        .body(
+            BaseResponse.ofSuccess(
+                SocialAccountSuccess.GET_SOCIAL_ACCOUNT_PLATFORM,
+                AuthResponse.SocialAccountPlatform.from(socialAccountPlatform)));
+  }
 
   @Override
   @PatchMapping
