+++ src/main/java/sopt/makers/authentication/application/auth/dto/response/AuthResponse.java
@@ -2,6 +2,7 @@ package sopt.makers.authentication.application.auth.dto.response;
 
 import static lombok.AccessLevel.PRIVATE;
 
+import sopt.makers.authentication.usecase.auth.port.in.GetSocialAccountUsecase;
 import sopt.makers.authentication.usecase.auth.port.in.VerifyPhoneVerificationUsecase;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -30,6 +31,13 @@ public final class AuthResponse {
   public record AuthenticateSocialAuthInfoForApp(String accessToken, String refreshToken) {
     public static AuthenticateSocialAuthInfoForApp of(String accessToken, String refreshToken) {
       return new AuthenticateSocialAuthInfoForApp(accessToken, refreshToken);
+    }
+  }
+
+  public record SocialAccountPlatform(@JsonProperty("platform") String platformName) {
+    public static SocialAccountPlatform from(
+        GetSocialAccountUsecase.SocialAccountPlatformInfo info) {
+      return new SocialAccountPlatform(info.platformName());
     }
   }
 }
