--- src/main/java/sopt/makers/authentication/application/user/api/UserApi.java
@@ -1,10 +1,0 @@
-package sopt.makers.authentication.application.user.api;
-
-import sopt.makers.authentication.support.common.api.BaseResponse;
-
-import org.springframework.http.ResponseEntity;
-
-public interface UserApi {
-
-  ResponseEntity<BaseResponse<?>> getRegisterSocialAccountPlatform(String name, String phone);
-}
