+++ src/main/java/sopt/makers/authentication/support/code/domain/success/AuthSuccess.java
@@ -15,6 +15,7 @@ public enum AuthSuccess implements SuccessCode {
   VERIFY_PHONE_VERIFICATION(HttpStatus.OK, "번호 인증에 성공했습니다."),
 
   CREATE_PHONE_VERIFICATION(HttpStatus.CREATED, "번호 인증 생성에 성공했습니다."),
+
   AUTHENTICATE_SOCIAL_ACCOUNT(HttpStatus.OK, "소셜 로그인에 성공했습니다.");
 
   private final HttpStatus status;
