+++ src/main/java/sopt/makers/authentication/support/code/domain/success/UserSuccess.java
@@ -12,7 +12,6 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor(access = PRIVATE)
 public enum UserSuccess implements SuccessCode {
-  GET_SOCIAL_ACCOUNT_PLATFORM(HttpStatus.OK, "가입 플랫폼 정보 조회에 성공했습니다."),
   ;
 
   private final HttpStatus status;
