+++ src/main/java/sopt/makers/authentication/usecase/auth/service/GetSocialAccountPlatformService.java
@@ -1,9 +1,9 @@
-package sopt.makers.authentication.usecase.user.service;
+package sopt.makers.authentication.usecase.auth.service;
 
 import sopt.makers.authentication.domain.auth.AuthPlatform;
 import sopt.makers.authentication.domain.user.User;
+import sopt.makers.authentication.usecase.auth.port.in.GetSocialAccountUsecase;
 import sopt.makers.authentication.usecase.auth.port.out.UserRepository;
-import sopt.makers.authentication.usecase.user.port.in.GetSocialAccountPlatform;
 
 import org.springframework.stereotype.Service;
 
@@ -11,7 +11,7 @@ import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
-public class GetSocialAccountPlatformService implements GetSocialAccountPlatform {
+public class GetSocialAccountPlatformService implements GetSocialAccountUsecase {
   private final UserRepository userRepository;
 
   @Override
